--- backend/domain/src/main/scala/io/mesazon/domain/gateway/OrganizationStage.scala
@@ -1,6 +1,0 @@
-package io.mesazon.domain.gateway
-
-enum OrganizationStage {
-  case DetailsProvided
-  case LogoProvided
-}
