+++ backend/domain/src/main/scala/io/mesazon/domain/gateway/gateway.scala
@@ -71,57 +71,3 @@ type AccessToken = AccessToken.T
 
 object Attempts extends RefinedType[Int, Positive]
 type Attempts = Attempts.T
-
-object OrganizationID extends RefinedTypeUUID
-type OrganizationID = OrganizationID.T
-
-object OrganizationLogoOriginalFileName extends RefinedType[String, NonEmptyTrimmed]
-type OrganizationLogoOriginalFileName = OrganizationLogoOriginalFileName.T
-
-object OrganizationLogoOriginalUrl extends RefinedType[String, NonEmptyTrimmedUnsafe]
-type OrganizationLogoOriginalUrl = OrganizationLogoOriginalUrl.T
-
-object OrganizationLogoNormalizedUrl extends RefinedType[String, NonEmptyTrimmedUnsafe]
-type OrganizationLogoNormalizedUrl = OrganizationLogoNormalizedUrl.T
-
-object OrganizationLogoOriginalBucketKey extends RefinedType[String, NonEmptyTrimmedUnsafe]
-type OrganizationLogoOriginalBucketKey = OrganizationLogoOriginalBucketKey.T
-
-object OrganizationLogoNormalizedBucketKey extends RefinedType[String, NonEmptyTrimmedUnsafe]
-type OrganizationLogoNormalizedBucketKey = OrganizationLogoNormalizedBucketKey.T
-
-object OrganizationName extends RefinedType[String, NonEmptyTrimmed]
-type OrganizationName = OrganizationName.T
-
-object OrganizationSlug extends RefinedType[String, SlugPredicate]
-type OrganizationSlug = OrganizationSlug.T
-
-object OrganizationTagline extends RefinedType[String, NonEmptyTrimmed]
-type OrganizationTagline = OrganizationTagline.T
-
-object OrganizationEmail extends RefinedType[String, EmailPredicate]
-type OrganizationEmail = OrganizationEmail.T
-
-object OrganizationPhoneNumber extends RefinedType[PhoneNumber, Pure]
-type OrganizationPhoneNumber = OrganizationPhoneNumber.T
-
-object OrganizationAddressLine1 extends RefinedType[String, NonEmptyTrimmed]
-type OrganizationAddressLine1 = OrganizationAddressLine1.T
-
-object OrganizationAddressLine2 extends RefinedType[String, NonEmptyTrimmed]
-type OrganizationAddressLine2 = OrganizationAddressLine2.T
-
-object OrganizationCity extends RefinedType[String, NonEmptyTrimmed]
-type OrganizationCity = OrganizationCity.T
-
-object OrganizationPostalCode extends RefinedType[String, NonEmptyTrimmed]
-type OrganizationPostalCode = OrganizationPostalCode.T
-
-object OrganizationCountry extends RefinedType[String, NonEmptyTrimmed]
-type OrganizationCountry = OrganizationCountry.T
-
-object OrganizationCompanyRegistrationNumber extends RefinedType[String, NonEmptyTrimmed]
-type OrganizationCompanyRegistrationNumber = OrganizationCompanyRegistrationNumber.T
-
-object OrganizationTaxID extends RefinedType[String, NonEmptyTrimmed]
-type OrganizationTaxID = OrganizationTaxID.T
