+++ docs-claude/adding-a-feature.md
@@ -52,6 +52,19 @@ Deriving the smithy arbitrary **from** the domain arbitrary is what makes the va
 
 Write `given arbInsertCustomerIndividual: Arbitrary[…]`, never an anonymous `given Arbitrary[…]`. Anonymous givens get a **synthesized name from the type**, and a singular/plural pair (`Arbitrary[AddCustomerBusinessContact]` vs `Arbitrary[AddCustomerBusinessContacts]`) synthesizes colliding names — the plural silently shadows the singular, which then fails to resolve with a misleading "no given instance … missing RefinedType.Mirror" error that survives a clean rebuild. Explicit names avoid it.
 
+## Consolidating an existing feature into this layout
+
+Older features left their pieces scattered (newtypes in `gateway.scala`, per-request `ServiceValidator` classes, arbitraries in the shared traits). Consolidating a feature `<Feature>` means moving every piece into the per-feature files above **without changing behavior**. Done for: `CustomerBook` (built this way), `OrganizationManagement`. The steps, in order:
+
+1. **Domain** — create `domain/gateway/<Feature>.scala` holding, in this order: the feature's newtypes (cut from `gateway.scala`), its enums (delete their one-enum files), its entry/value case classes, its request case classes (delete their files). Everything stays in package `io.mesazon.domain.gateway`, so **no call site changes** — this step is pure file moves.
+2. **Validator** — replace the per-request `<Request>ServiceValidator` classes with one `validation/service/<Feature>RequestValidator.scala`: a plain class (not a `ServiceValidator`) with one public `validated<Request>PostRequest(request): IO[ValidationError, <Domain>]` per fallible request, each delegating to a private `UIO[ValidatedNec[InvalidFieldError, <Domain>]]` via `toValidatedRequestIO`. Callers change from `validator.validate(r)` to `validator.validated<Request>PostRequest(r)`; update the service, `Main`'s layer list, and any spec `provide`.
+3. **Domain arbitraries** — create `<Feature>DomainArbitraries` in test-kit and move every feature `Arbitrary` (enums included) out of `GatewayArbitraries` into it, naming each given (`arb<Type>`). Generic helpers shared across features (e.g. `genEntriesWithSingleDefault`) stay `protected` in `GatewayArbitraries`.
+4. **Smithy arbitraries** — create `<Feature>SmithyArbitraries` in gateway-core test utils and move the feature's `Transformer`s and smithy-request arbitraries out of the shared `SmithyArbitraries`.
+5. **Re-wire mixins** — every spec that used a moved arbitrary now mixes in the feature trait instead of (or in addition to) the shared one. Watch the two indirect consumers: `RepositoryArbitraries` (row arbitraries `Gen.resultOf` a row whose fields need the feature's arbitraries — extend the feature's domain trait) and any spec that only used the shared trait *for* this feature's givens (swap the mixin).
+6. **Rename specs** — the validator spec becomes `<Feature>RequestValidatorSpec` mixing `<Feature>SmithyArbitraries`.
+7. **Docs** — per the [Rename rule](../CLAUDE.md): grep `docs-claude/` for every renamed identifier (old validator class names above all) and update; update the feature doc's key-files section.
+8. **Verify** — full `Test/compile`, then the whole `gateway-core` and `gateway-it` suites. A consolidation PR must be behavior-neutral: no smithy, SQL, or assertion changes.
+
 ## Order of work
 
 1. **Smithy** contract (`smithy4sCodegen` must be green).
